requirements.txt
ccxt>=4.0
aiohttp>=3.9
pyyaml>=6.0
aiosqlite>=0.19
flask>=3.0
pandas>=2.0
numpy>=1.24
ta>=0.11
python-dotenv>=1.0
backtrader>=1.9.78.123
pytest>=8.0
pytest-asyncio>=0.23
